requirements.txt
torch
torchvision
datasets
openai
streamlit
python-dotenv
tenacity
pillow
pandas
